Reorganized all files into appropriate folders as well as refactored imports for each files.
