+++ cse442_App/android/local.properties
@@ -1,5 +1,5 @@
-sdk.dir=/Users/yingweili/Library/Android/sdk
-flutter.sdk=/Users/yingweili/flutter
+sdk.dir=C:\\Users\\Brand\\AppData\\Local\\Android\\Sdk
+flutter.sdk=C:\\Users\\Brand\\flutter
 =======
 flutter.buildMode=debug
 flutter.versionName=1.0.0
